Paired with bd-socketio app.
Added node module fetch.
Added command to upload images from competitions to my Multiplayer App for reviewing them. Videos don't work, only images.
Changed a few other bits.

+++ index.js
@@ -1,29 +1,30 @@
+const fetch = require("node-fetch");
 const Discord = require('discord.js');
 const client = new Discord.Client();
-const data = require("./data.json");
-const version = "v5.2";
+const Data = require("./data.json");
+const version = "v6.0";
 
 const imageChannels = [];
 const videoChannels = [];
-data.servers.forEach( s => {
+Data.servers.forEach(s => {
     imageChannels.push(s.imageChannel.id);
 });
-data.servers.forEach( s => {
+Data.servers.forEach(s => {
     videoChannels.push(s.videoChannel.id);
 });
 
 
-client.on('ready', () => {console.log(`Logged in as ${client.user.tag}!`)});
+client.on('ready', () => { console.log(`Logged in as ${client.user.tag}!`) });
 client.on('error', err => { console.log(err) });
 
 client.on('message', msg => {
     if (msg.author.bot) return;
-    
+
     if (msg.isMemberMentioned(client.user)) {
         checkCmd(msg);
     }
     if (imageChannels.includes(msg.channel.id)) {
-        let timeWait =  setTimeout(checkMsg, 1000); //wait until discord auto embeds video
+        let timeWait = setTimeout(checkMsg, 1000); //wait until discord auto embeds video
         function checkMsg() {
             if (isImage(msg)) { // if msg is a image in the image submission channel
                 cleanChannel(msg, 'IMAGE');
@@ -40,8 +41,8 @@ client.on('message', msg => {
 })
 
 function cleanChannel(msg, type, cmd) { // removes all none submission messages from the submission channel
-    if (type == 'VIDEO') { 
-        msg.channel.fetchMessages({ limit: 100 }).then(messages => { 
+    if (type == 'VIDEO') {
+        msg.channel.fetchMessages({ limit: 100 }).then(messages => {
             messages.forEach(message => {
                 if (!isVideo(message)) { // delete all non-video messages
                     deleteMsg(message, 0);
@@ -50,8 +51,8 @@ function cleanChannel(msg, type, cmd) { // removes all none submission messages 
                 }
             });
         });
-    } else if (type == 'IMAGE') { 
-        msg.channel.fetchMessages({ limit: 100 }).then(messages => { 
+    } else if (type == 'IMAGE') {
+        msg.channel.fetchMessages({ limit: 100 }).then(messages => {
             messages.forEach(message => {
                 if (!isImage(message)) { // delete all non-image messages
                     deleteMsg(message, 0);
@@ -63,7 +64,7 @@ function cleanChannel(msg, type, cmd) { // removes all none submission messages 
     };
 }
 
-function isVideo (msg) {
+function isVideo(msg) {
     if (msg.embeds[0]) {
         if (msg.embeds[0].type == "video") {
             return true;
@@ -77,7 +78,7 @@ function isVideo (msg) {
     }
 }
 
-function isImage (msg) {
+function isImage(msg) {
     if (msg.embeds[0]) {
         if (msg.embeds[0].type == 'image') {
             return true;
@@ -97,6 +98,7 @@ function checkCmd(msg) {
     var reset = str.search('RESET');
     var clean = str.search('CLEAN');
     var help = str.search('HELP');
+    var upload = str.search('UPLOAD');
 
     if (hi > -1) {
         sendMsg(msg, `Hello, ${msg.author}`, -1);
@@ -110,19 +112,90 @@ function checkCmd(msg) {
         } else {
             console.log(`Clean Command does not work in channel with id: ${msg.channel.id}`)
         }
-    } else if (reset > -1) {
-        if (videoChannels.includes(msg.channel.id) || imageChannels.includes(msg.channel.id)){
-            data.mods.forEach(m => {
-                if (m.id == msg.author.id) {
-                    deleteAll(msg);
-                }
-            })
+    } else if (upload > -1 && isAdmin(msg)) {
+        uploadSubmissions(msg); //uploads submissions to external app
+    } else if (reset > -1 && isAdmin(msg)) {
+        if (videoChannels.includes(msg.channel.id) || imageChannels.includes(msg.channel.id)) {
+            deleteAll(msg);
         }
     }
 }
 
+function isAdmin(msg) {
+    if (Data.admins.find(a => a.id === msg.author.id)) {
+        return true
+    }
+}
+
+function uploadSubmissions(msg) {
+    console.log("attemping to upload data");
+    if (imageChannels.includes(msg.channel.id)) {
+        uploadImages(msg);
+    } else if (videoChannels.includes(msg.channel.id)) {
+        uploadVideos(msg);
+    } else {
+        console.log(`Upload Command does not work in channel with id: ${msg.channel.id}`)
+    }
+
+    function uploadImages(msg) {
+        let images = [];
+        msg.channel.fetchMessages({ limit: 100 }).then(messages => {
+            messages.forEach(message => {
+                if (isImage(message)) {
+                    let image = getUrl(message);
+                    let name = message.author.username;
+                    images.push({ url: image, text: name });
+                }
+            });
+            uploadData(images);
+        });
+
+    }
+
+    function uploadVideos() {
+        let videos = [];
+        msg.channel.fetchMessages({ limit: 100 }).then(messages => {
+            messages.forEach(message => {
+                if (isVideo(message)) {
+                    let image = getUrl(message);
+                    let name = message.author.username;
+                    images.push({ url: image, text: name });
+                }
+            });
+            uploadData(videos);
+        });
+    }
+
+    function getUrl(msg) {
+        if (msg.attachments.size > 0) {
+            let attachArray = Array.from(msg.attachments.values());
+            let url = attachArray[0].url;
+            return url;
+        }
+    }
+
+    function uploadData(data) {
+        console.log(data.length);
+        // uploads images to my multiplayer veiwing app.
+        fetch('https://bd-socketio.glitch.me/', {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify(data),
+        })
+            .then(response => response.text())
+            .then(data => {
+                console.log('Success:', data);
+            })
+            .catch((error) => {
+                console.error('Error:', error);
+            });
+    }
+}
+
 function deleteAll(msg) {
-    data.servers.forEach(server => {
+    Data.servers.forEach(server => {
         if (server.id == msg.guild.id) {
             if (msg.channel.id == server.videoChannel.id || msg.channel.id == server.imageChannel.id) {
                 msg.channel.fetchMessages({ limit: 100 }).then(messages => {
@@ -138,7 +211,7 @@ function deleteAll(msg) {
 }
 
 function deleteMsg(msg, delay) {
-    data.servers.forEach(server => {
+    Data.servers.forEach(server => {
         if (server.id == msg.guild.id) {
             msg.delete(delay).catch(err => {
                 console.log(`Might be Missing MANAGE_MESSAGES permissions to delete ${msg.author.username}'s message`);
